+++ src/Proto.Actor/DependencyInjection/DIExtension.cs
@@ -27,6 +27,6 @@ namespace Proto.DependencyInjection
             return actorSystem;
         }
 
-        public static IDependencyResolver DI(this ActorSystem system) => system.Extensions.Get<DIExtension>().Resolver;
+        public static IDependencyResolver DI(this ActorSystem system) => system.Extensions.Get<DIExtension>()!.Resolver;
     }
 }
