+++ src/Proto.Remote.GrpcNet/GrpcNetRemote.cs
@@ -26,7 +26,7 @@ namespace Proto.Remote.GrpcNet
         {
             System = system;
             _config = config;
-            System.Extensions.Register(this);
+            System.Extensions.Register(new RemoteExtension(this));
             System.Extensions.Register(config.Serialization);
         }
