+++ src/Proto.Remote/IRemote.cs
@@ -3,7 +3,15 @@ using Proto.Extensions;
 
 namespace Proto.Remote
 {
-    public interface IRemote : IActorSystemExtension<IRemote>
+    public class RemoteExtension : IActorSystemExtension<RemoteExtension>
+    {
+        public IRemote Remote { get; }
+        public RemoteExtension(IRemote remote)
+        {
+            Remote = remote;
+        }
+    }
+    public interface IRemote 
     {
         RemoteConfigBase Config { get; }
         ActorSystem System { get; }
