+++ src/Proto.Remote/IRemoteExtensions.cs
@@ -5,7 +5,7 @@ namespace Proto.Remote
 {
     public static class IRemoteExtensions
     {
-        public static IRemote Remote(this ActorSystem system) => system.Extensions.Get<IRemote>();
+        public static IRemote Remote(this ActorSystem system) => system.Extensions.Get<RemoteExtension>()!.Remote;
 
         public static IRemote Remote(this IContext context) => context.System.Remote();
